basic dashboard page w/ routing
